+++ src/azure-cli/azure/cli/command_modules/identity/_help.py
@@ -1,0 +1,21 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from knack.help_files import helps
+
+helps['identity'] = """
+type: group
+short-summary: Managed Service Identities
+"""
+
+helps['identity list'] = """
+type: command
+short-summary: List Managed Service Identities
+"""
+
+helps['identity list-operations'] = """
+type: command
+short-summary: Lists available operations for the Managed Identity provider
+"""
