+++ src/azure-cli/azure/cli/command_modules/identity/_params.py
@@ -1,0 +1,21 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from knack.arguments import CLIArgumentType
+
+from azure.cli.core.commands.parameters import get_location_type, tags_type
+
+
+name_arg_type = CLIArgumentType(options_list=('--name', '-n'), metavar='NAME')
+
+
+def load_arguments(self, _):
+
+    with self.argument_context('identity') as c:
+        c.argument('resource_name', arg_type=name_arg_type, id_part='name')
+
+    with self.argument_context('identity create') as c:
+        c.argument('location', get_location_type(self.cli_ctx))
+        c.argument('tags', tags_type)
