+++ src/azure-cli/azure/cli/command_modules/identity/_validators.py
@@ -1,0 +1,11 @@
+# --------------------------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for license information.
+# --------------------------------------------------------------------------------------------
+
+from azure.cli.core.commands.validators import get_default_location_from_resource_group, validate_tags
+
+
+def process_msi_namespace(cmd, namespace):
+    get_default_location_from_resource_group(cmd, namespace)
+    validate_tags(namespace)
