+++ src/azure-cli/azure/cli/command_modules/role/_client_factory.py
@@ -27,17 +27,3 @@ def _graph_client_factory(cli_ctx, **_):
                                        base_url=cli_ctx.cloud.endpoints.active_directory_graph_resource_id)
     configure_common_settings(cli_ctx, client)
     return client
-
-
-def _msi_client_factory(cli_ctx, **_):
-    from azure.mgmt.msi import ManagedServiceIdentityClient
-    from azure.cli.core.commands.client_factory import get_mgmt_service_client
-    return get_mgmt_service_client(cli_ctx, ManagedServiceIdentityClient)
-
-
-def _msi_user_identities_operations(cli_ctx, _):
-    return _msi_client_factory(cli_ctx).user_assigned_identities
-
-
-def _msi_operations_operations(cli_ctx, _):
-    return _msi_client_factory(cli_ctx).operations
