+++ src/azure-cli/azure/cli/command_modules/role/_validators.py
@@ -113,8 +113,3 @@ def process_assignment_namespace(cmd, namespace):  # pylint: disable=unused-argu
     resource_group = namespace.resource_group_name
     if namespace.scope and resource_group and getattr(resource_group, 'is_default', None):
         namespace.resource_group_name = None  # drop configured defaults
-
-
-def process_msi_namespace(cmd, namespace):
-    get_default_location_from_resource_group(cmd, namespace)
-    validate_tags(namespace)
