+++ src/azure-cli/azure/cli/command_modules/role/custom.py
@@ -1874,11 +1874,3 @@ def _random_password(length):
 
     password = first_character + ''.join(password_list)
     return password
-
-
-def list_user_assigned_identities(cmd, resource_group_name=None):
-    from azure.cli.command_modules.role._client_factory import _msi_client_factory
-    client = _msi_client_factory(cmd.cli_ctx)
-    if resource_group_name:
-        return client.user_assigned_identities.list_by_resource_group(resource_group_name)
-    return client.user_assigned_identities.list_by_subscription()
